Refactor logging_config into smaller modules
